requirements.txt
gdown
tensorflow
numpy
Pillow
streamlit
python-dotenv
openai
